move auth property iterator declaration into auth_context.h
